add updating weekly calendar of recipes

+++ WeeklyBite - back/src/main/java/com/backend/weeklybite/repository/IngredientRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 public interface IngredientRepository extends JpaRepository<Ingredient, Long>, JpaSpecificationExecutor<Ingredient> {
-    Ingredient findByNameAndQuantityAndUnit(String name, Integer quantity, String unit);
+    Ingredient findByNameAndQuantityAndUnit(String name, Double quantity, String unit);
 }
